change DaharNavBar in home.dart to NavBar and move it to a separate file
