+++ lib/screens/home.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:dahar/global_styles.dart';
+import 'package:dahar/components/navbar.dart';
 
 class Home extends StatelessWidget {
   const Home({Key? key}) : super(key: key);
@@ -16,7 +17,7 @@ class Home extends StatelessWidget {
             children: const [Popular(), Closest()],
           ),
         ),
-        bottomNavigationBar: const DaharNavBar());
+        bottomNavigationBar: const NavBar());
   }
 }
 
@@ -64,68 +65,6 @@ class _DaharAppBarState extends State<DaharAppBar> {
             )
           ],
         ),
-      ),
-    );
-  }
-}
-
-class DaharNavBar extends StatefulWidget {
-  const DaharNavBar({Key? key}) : super(key: key);
-
-  @override
-  State<DaharNavBar> createState() => _DaharNavBarState();
-}
-
-class _DaharNavBarState extends State<DaharNavBar> {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: BoxDecoration(color: Colors.white, boxShadow: [
-        BoxShadow(
-            offset: const Offset(0, 20),
-            blurRadius: 90,
-            spreadRadius: 0,
-            color: Colors.black.withOpacity(0.25))
-      ]),
-      padding: const EdgeInsets.only(top: 10, bottom: 10),
-      // color: Colors.blue,
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          IconButton(
-            alignment: Alignment.center,
-            icon: const Icon(
-              Icons.home_filled,
-            ),
-            color: color1,
-            onPressed: () {},
-          ),
-          IconButton(
-            alignment: Alignment.center,
-            icon: const Icon(
-              Icons.favorite_border,
-            ),
-            color: color1,
-            onPressed: () {},
-          ),
-          IconButton(
-            alignment: Alignment.center,
-            icon: const Icon(
-              Icons.shopping_cart_outlined,
-            ),
-            color: color1,
-            onPressed: () {},
-          ),
-          IconButton(
-            alignment: Alignment.center,
-            icon: const Icon(
-              Icons.perm_identity_outlined,
-            ),
-            color: color1,
-            onPressed: () {},
-          ),
-        ],
       ),
     );
   }
