+++ lib/screens/order_history.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:dahar/global_styles.dart';
+import 'package:dahar/components/navbar.dart';
 
 class OrderHistory extends StatelessWidget {
   const OrderHistory({Key? key}) : super(key: key);
@@ -7,24 +8,24 @@ class OrderHistory extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: const PreferredSize(
-          preferredSize: Size.fromHeight(100), child: HistoryAppBar()),
-      body: Container(
-        padding: const EdgeInsets.symmetric(horizontal: 25),
-        color: Colors.white,
-        child: ListView(children: const [
-          HistoryItem(
-            isConfirmed: false,
-          ),
-          HistoryItem(
-            isConfirmed: false,
-          ),
-          HistoryItem(
-            isConfirmed: false,
-          )
-        ]),
-      ),
-    );
+        appBar: const PreferredSize(
+            preferredSize: Size.fromHeight(100), child: HistoryAppBar()),
+        body: Container(
+          padding: const EdgeInsets.symmetric(horizontal: 25),
+          color: Colors.white,
+          child: ListView(children: const [
+            HistoryItem(
+              isConfirmed: false,
+            ),
+            HistoryItem(
+              isConfirmed: false,
+            ),
+            HistoryItem(
+              isConfirmed: false,
+            )
+          ]),
+        ),
+        bottomNavigationBar: const NavBar());
   }
 }
